tests: create separate test file for torch and its libraries
Create a separate `test_pytorch.py` test for `torch` and its
associated libraries. Move existing tests to the new file.
